fix issues to pass energy tests for inverse_power_stilinger_cut_quad

+++ source/pele/inversepower_stillinger_cut_quad.hpp
@@ -32,7 +32,7 @@ struct InversePowerStillingerQuadCutInteraction : BaseInteraction {
   const double c2;               // quadratic term in cutoff contribution
   const double c4;               // quartic term in cutoff contribution
   const double m_v0;             // overall energy scale for the full potential
-  InversePowerStillingerQuadCutInteraction(const size_t pow, const double v0,
+  InversePowerStillingerQuadCutInteraction(const int pow, const double v0,
                                            const double cutoff_factor)
       : m_pow(pow), m_pow_by_2(pow / 2), m_v0(v0),
         m_cutoff_factor(cutoff_factor),
